tests: kernel: events: merge sys_event into event_api
The sys_event test suite duplicated the event_api coverage with a
simpler subset of the same k_event_* checks, differing only in that it
enabled userspace testing. Drop the redundant suite and the sys_event
terminology, and fold the userspace coverage into event_api instead.
