+++ tests/kernel/events/event_api/prj.conf
@@ -1,3 +1,4 @@
 CONFIG_ZTEST=y
 CONFIG_IRQ_OFFLOAD=y
+CONFIG_TEST_USERSPACE=y
 CONFIG_EVENTS=y
