--- tests/kernel/events/sys_event/prj.conf
@@ -1,4 +1,0 @@
-CONFIG_ZTEST=y
-CONFIG_TEST_USERSPACE=y
-CONFIG_ZTEST_FATAL_HOOK=y
-CONFIG_EVENTS=y
