--- tests/kernel/events/sys_event/testcase.yaml
@@ -1,9 +1,0 @@
-tests:
-  kernel.events.usage:
-    filter: CONFIG_ARCH_HAS_USERSPACE
-    arch_exclude:
-      - posix
-    tags:
-      - kernel
-      - userspace
-    ignore_faults: true
